+++ public/index.php
@@ -12,8 +12,8 @@ if (PHP_SAPI == 'cli-server') {
 error_reporting(E_ALL);
 //ini_set('display_errors', TRUE);
 //ini_set('display_startup_errors', TRUE);
-define('FILE_SHARING_VERSION', '3.10.0');
-define('FILE_SHARING_TIMESTAMP', '2022-11-13T11:30:00+01:00');
+define('FILE_SHARING_VERSION', '3.11.0-dev');
+define('FILE_SHARING_TIMESTAMP', '2022-12-09T18:30:00+01:00');
 
 require __DIR__ . '/../vendor/autoload.php';