+++ sql/all-records.sql
@@ -1,3 +1,8 @@
+INSERT INTO `imhhfs_access` (`id`, `name`, `icon`, `button`, `deleted`, `hidden`, `updated_at`, `created_at`) VALUES
+(1,	'private',	'lock',	'success',	0,	0,	now(),	now()),
+(2,	'shareable',	'link',	'warning',	0,	0,	now(),	now()),
+(3,	'public',	'eye',	'danger',	0,	0,	now(),	now());
+
 INSERT INTO `imhhfs_role` (`id`, `name`, `deleted`, `hidden`, `updated_at`, `created_at`) VALUES
 (1, 'guest', 0, 0, now(), now()),
 (2, 'member', 0, 0, now(), now()),