+++ sql/upload_limit-records.sql
@@ -1,0 +1,2 @@
+INSERT INTO `imhhfs_upload_limit` (`id`, `name`, `size`, `files`, `folders`, `deleted`, `hidden`, `updated_at`, `created_at`) VALUES
+(1, 'general', '104857600', 250, 250, 0, 0, now(), now());