+++ src/MappedSuperclass/Base.php
@@ -2,6 +2,7 @@
 namespace App\MappedSuperclass;
 
 use Doctrine\ORM\Mapping as ORM;
+use Ramsey\Uuid\Uuid;
 
 /**
  * This class is used for Entities
@@ -64,6 +65,7 @@ class Base {
      * Set $deleted
      * 
      * @param boolean $deleted
+     * @return Base
      */
     public function setDeleted($deleted) {
         $this->deleted = $deleted;
@@ -84,6 +86,7 @@ class Base {
      * Set $hidden
      * 
      * @param boolean $hidden
+     * @return Base
      */
     public function setHidden($hidden) {
         $this->hidden = $hidden;
@@ -104,6 +107,7 @@ class Base {
      * Set $updatedAt
      * 
      * @param \DateTime $updatedAt
+     * @return Base
      */
     public function setUpdatedAt($updatedAt) {
         $this->updatedAt = $updatedAt;
@@ -124,9 +128,21 @@ class Base {
      * Set $createdAt
      * 
      * @param \DateTime $createdAt
+     * @return Base
      */
     public function setCreatedAt($createdAt) {
         $this->createdAt = $createdAt;
+        
+        return $this;
+    }
+    
+    /**
+     * Set $id with Ramsey/Uuid
+     * 
+     * @return Base
+     */
+    protected function setIdWithRamsey() {
+        $this->id = Uuid::uuid4();
         
         return $this;
     }