+++ src/dependencies.php
@@ -74,7 +74,7 @@ $container['view'] = function ($c) {
     $view->addExtension(new App\Twig\FileExtension($c));
     $view->addExtension(new App\Twig\GeneralExtension($c));
     $view->addExtension(new App\Twig\LanguageExtension($c, $router, $uri));
-    $view->addExtension(new nochso\HtmlCompressTwig\Extension());
+    $view->addExtension(new nochso\HtmlCompressTwig\Extension($settings['renderer']['htmlcompress']));
     $view->addExtension(new Slim\Views\TwigExtension($router, $uri));
     
     if ($settings['renderer']['debug']) {